pyproject.toml
[tool.poetry]
name = "couswork-django"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
flake8 = "^7.1.2"
black = "^25.1.0"
django-phonenumber-field = "^8.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
